ABOUT US Page Features and Layout
Added a new feature - filters
Added filtering of the user's clothes by type/color/occasion
The filter works on the My Clothes page and also if the user is searching for a product in the search bar.
Arranged the ABOUTUS page - Added a caption to the page about our site and I also added that the user will have the option to write a review of the site and the review will be sent directly to the company's email.